+++ site/components/Components.jsx
@@ -11,7 +11,7 @@ import FaIcon from "faicon/FaIcon";
 
 
 
-export default class Showcase extends ShallowComponent {
+export default class Components extends ShallowComponent {
 
     constructor(props: Object) {
         super(props);